poc3/backend/requirements.txt

flask
openai>=1.3.7
python-dotenv

# Jira Integration
jira>=3.5.0   
